chore(validation): birfurcated rules
