+++ packages/validation/src/index.ts
@@ -1,6 +1,8 @@
+export * from './configuration';
+export * from './rule-provider';
+export * from './rules';
+export * from './subscribers/validation-errors-custom-attribute';
 export * from './validate-binding-behavior';
 export * from './validation-controller';
 export * from './validator';
-export * from './rule';
-export * from './configuration';
-export * from './subscribers/validation-errors-custom-attribute';
+
