+++ packages/validation/src/validate-binding-behavior.ts
@@ -14,7 +14,7 @@ import {
   IBinding,
   BindingBehaviorExpression,
 } from '@aurelia/runtime';
-import { PropertyRule } from './rule';
+import { PropertyRule } from './rule-provider';
 import { BindingWithBehavior, IValidationController, ValidationController, BindingInfo } from './validation-controller';
 
 /**
