+++ packages/validation/src/validation-controller.ts
@@ -18,14 +18,13 @@ import {
   State
 } from '@aurelia/runtime';
 import {
-  IValidateable,
   parsePropertyName,
   PropertyAccessor,
   PropertyRule,
   ValidationResult,
-  BaseValidationRule
-} from './rule';
+} from './rule-provider';
 import { IValidator } from './validator';
+import { IValidateable, BaseValidationRule } from './rules';
 
 export const VALIDATION_EVENT_CHANNEL = 'au:validation';
