+++ packages/validation/src/validation-customization-options.ts
@@ -1,4 +1,4 @@
-import { ICustomMessage } from './rule';
+import { ICustomMessage } from './rule-provider';
 import { IValidator } from './validator';
 import { Class } from '@aurelia/kernel';
 import { ValidationTrigger } from './validate-binding-behavior';
